added, deleted some comments

+++ app/controllers/usercontroller.php
@@ -17,7 +17,12 @@ class UserController extends Controller
         $this->userService = new UserService();
     }
 
-    public function getAll(){
+    public function getAll(){ //used while an admin views the Users list
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
+        
         $offset = NULL;
         $limit = NULL;
 
@@ -33,7 +38,7 @@ class UserController extends Controller
         $this->respond($vets);
     }
 
-    public function getByEmail($email)
+    public function getByEmail($email) //used for logging in.
     {
         $user = $this->userService->getByEmail($email);
 
@@ -45,8 +50,13 @@ class UserController extends Controller
         $this->respond($user);
     }
 
-    public function getById($id)
+    public function getById($id) //used while editing a user
     {
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
+
         $user = $this->userService->getById($id);
 
         if (!$user) {
@@ -57,7 +67,7 @@ class UserController extends Controller
         $this->respond($user);
     }
 
-    public function create()
+    public function create() //used while registering
     {
         try {
             $requestBody = file_get_contents('php://input');
@@ -77,7 +87,11 @@ class UserController extends Controller
         $this->respond($newUser);
     }
 
-    public function update($id) {
+    public function update($id) { //used while updating the user
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
         try {
             $requestBody = file_get_contents('php://input');
             $userData = json_decode($requestBody);
@@ -96,8 +110,12 @@ class UserController extends Controller
         $this->respond($userToUpdate);
     }
 
-    public function delete($id)
+    public function delete($id) //used while deleting the user
     {
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
         try {
             $this->userService->delete($id);
         } catch (Exception $e) {
@@ -108,39 +126,31 @@ class UserController extends Controller
     }
 
     public function login() {
-
-        // read user data from request body
+        
         $postedUser = $this->createObjectFromPostedJson("Models\\User");
 
-        // get user from db
         $user = $this->userService->checkUsernamePassword($postedUser->getEmail(), $postedUser->getPassword());
 
-        // if the method returned false, the username and/or password were incorrect
         if(!$user) {
             $this->respondWithError(401, "Invalid login");
             return;
         }
 
-        // generate jwt
         $tokenResponse = $this->generateJwt($user);       
 
         $this->respond($tokenResponse);    
     }
 
     public function generateJwt($user) {
-        $secret_key = "YOUR_SECRET_KEY";
+        $secret_key = "megasuperamazinglysecurekey";
 
-        $issuer = "THE_ISSUER"; // this can be the domain/servername that issues the token
-        $audience = "THE_AUDIENCE"; // this can be the domain/servername that checks the token
+        $issuer = "THE_ISSUER"; 
+        $audience = "THE_AUDIENCE"; 
 
         $issuedAt = time(); // issued at
         $notbefore = $issuedAt; //not valid before 
-        $expire = $issuedAt + 600; // expiration time is set at +600 seconds (10 minutes)
+        $expire = $issuedAt + 600; //(10 minutes)
 
-        // JWT expiration times should be kept short (10-30 minutes)
-        // A refresh token system should be implemented if we want clients to stay logged in for longer periods
-
-        // note how these claims are 3 characters long to keep the JWT as small as possible
         $payload = array(
             "iss" => $issuer,
             "aud" => $audience,
@@ -160,6 +170,7 @@ class UserController extends Controller
                 "message" => "Successful login.",
                 "jwt" => $jwt,
                 "email" => $user->getEmail(),
+                "role" => $user->getRole(),
                 "expireAt" => $expire
             );
     }
