+++ app/controllers/vetController.php
@@ -13,7 +13,9 @@ class VetController extends Controller{
         $this->vetService = new VetService();
     }
 
-    public function getAll(){
+    public function getAll(){ //used in homepage while showing the vets
+
+        //im not checking for token because visitors (not logged in) should also see the vets.
         $offset = NULL;
         $limit = NULL;
 
@@ -29,8 +31,12 @@ class VetController extends Controller{
         $this->respond($vets);
     }
 
-    public function getById($id)
+    public function getById($id) //used while editing the vet
     {
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
         $vet = $this->vetService->getById($id);
 
         if (!$vet) {
@@ -41,8 +47,13 @@ class VetController extends Controller{
         $this->respond($vet);
     }
 
-    public function create()
+    public function create() //used while adding the vet 
     {
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
+
         try {
             $requestBody = file_get_contents('php://input');
             $vetData = json_decode($requestBody);
@@ -62,7 +73,11 @@ class VetController extends Controller{
         $this->respond($newVet);
     }
 
-    public function update($id) {
+    public function update($id) { //used while updating the vet info
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
         try {
             $requestBody = file_get_contents('php://input');
             $vetData = json_decode($requestBody);
@@ -81,8 +96,12 @@ class VetController extends Controller{
         $this->respond($vetToUpdate);
     }
 
-    public function delete($id)
+    public function delete($id) //used while deleting the vet from the system
     {
+        $admin = $this->checkForAdmin();
+        if (!$admin) {
+            return;
+        }
         try {
             $this->vetService->delete($id);
         } catch (Exception $e) {
